fix: run browser demo from monorepo
