+++ src/packages/browser-demo/package.json
@@ -5,37 +5,40 @@
   "description": "A browser Lynxtron application",
   "main": "main.js",
   "scripts": {
-    "dev": "cross-env TARGET_ENV=desktop NODE_ENV=development concurrently -k --prefix none \"rspeedy dev\" \"dev-ready-rspeedy && rspack dev\"",
+    "dev": "node ./scripts/postinstall.cjs && cross-env TARGET_ENV=desktop NODE_ENV=development concurrently -k --prefix none \"rspeedy dev\" \"dev-ready-rspeedy && rspack dev\"",
     "dev:web": "cross-env TARGET_ENV=web NODE_ENV=development concurrently -k --prefix node \"rspeedy dev\" \"rspack serve\"",
     "build": "rspeedy build && rspack build",
-    "start": "cross-env TARGET_ENV=desktop npm run build && lynxtron ./dist/desktop",
+    "start": "node ./scripts/postinstall.cjs && cross-env TARGET_ENV=desktop npm run build && lynxtron ./dist/desktop",
     "start:web": "cross-env TARGET_ENV=web npm run build && npx serve ./dist/web",
     "test": "vitest run",
     "pack": "npm run build && lynxtron-builder --publish never",
-    "postinstall": "node ./scripts/postinstall.js"
+    "postinstall": "node ./scripts/postinstall.cjs"
   },
   "private": true,
   "devDependencies": {
+    "@lynx-js/cef-webview": "0.0.10",
     "@lynx-js/lynx-core": "latest",
     "@lynx-js/lynxtron": "workspace:*",
     "@lynx-js/lynxtron-builder": "workspace:*",
     "@lynx-js/lynxtron-dev-plugins": "workspace:*",
-    "@lynx-js/preact-devtools": "^5.0.1-c0c936b",
-    "@lynx-js/react": "0.115.4",
-    "@lynx-js/react-rsbuild-plugin": "^0.12.5",
-    "@lynx-js/rspeedy": "^0.13.0",
-    "@lynx-js/types": "3.6.0",
+    "@lynx-js/preact-devtools": "latest",
+    "@lynx-js/react": "latest",
+    "@lynx-js/react-rsbuild-plugin": "latest",
+    "@lynx-js/rspeedy": "latest",
+    "@lynx-js/types": "latest",
     "@lynx-js/web-core": "latest",
     "@lynx-js/web-elements": "latest",
-    "@lynx-js/cef-webview": "0.0.10",
-    "@rsbuild/plugin-type-check": "1.3.1",
+    "@rsbuild/core": "2.1.5",
+    "@rsbuild/plugin-type-check": "latest",
     "@rspack/cli": "^1.7.5",
     "@rspack/core": "^1.7.5",
     "@testing-library/jest-dom": "^6.9.0",
+    "@types/react": "^18.3.24",
     "compressing": "^1.10.0",
     "concurrently": "^8.2.2",
     "cross-env": "^10.1.0",
     "jsdom": "^26.1.0",
+    "tslib": "latest",
     "typescript": "~5.9.3",
     "vitest": "^3.2.4"
   },
