+++ src/packages/browser-demo/src/app/__tests__/index.test.tsx
@@ -2,42 +2,14 @@
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
 import '@testing-library/jest-dom';
-import { expect, test, vi } from 'vitest';
+import { expect, test } from 'vitest';
 import { render, getQueriesForElement } from '@lynx-js/react/testing-library';
 
 import { App } from '../App.jsx';
 
 test('App', async () => {
-  const cb = vi.fn();
-
-  render(
-    <App
-      onRender={() => {
-        cb(`__MAIN_THREAD__: ${__MAIN_THREAD__}`);
-      }}
-    />
-  );
-  expect(elementTree.root).toMatchInlineSnapshot(`
-    <page>
-      <view
-        class="Background"
-      >
-        <text
-          class="Content"
-        >
-          Hello, Lynxtron!
-        </text>
-      </view>
-    </page>
-  `);
+  render(<App />);
   const { findByText } = getQueriesForElement(elementTree.root!);
-  const element = await findByText('Hello, Lynxtron!');
-  expect(element).toBeInTheDocument();
-  expect(element).toMatchInlineSnapshot(`
-    <text
-      class="Content"
-    >
-      Hello, Lynxtron!
-    </text>
-  `);
+  expect(await findByText('Lynxtron')).toBeInTheDocument();
+  expect(await findByText('Quick Access')).toBeInTheDocument();
 });
