+++ src/packages/browser-demo/vitest.config.ts
@@ -1,9 +1,11 @@
 import { defineConfig, mergeConfig } from 'vitest/config';
 import { createVitestConfig } from '@lynx-js/react/testing-library/vitest-config';
 
-const defaultConfig = await createVitestConfig();
 const config = defineConfig({
   test: {},
 });
 
-export default mergeConfig(defaultConfig, config);
+export default async function createConfig() {
+  const defaultConfig = await createVitestConfig();
+  return mergeConfig(defaultConfig, config);
+}
